Create Otu Julia ujunwa 2024714031
This Python program implements a grading system using 'if' and 'elif' statements. It takes a student's score as input and print the corresponding grade base on the following scale:
• 70 - 100: Grade A
• 60 - 69: Grade B
• 50 - 59: Grade C
• 45 - 49: Grade D
• 40 - 44: Grade E
• 20 - 29: Grade F

For each valid score, the program also prints "YES" to confirm the grading was successful. If the score is outside the valid range, it prints 'invalid score' and 'NO'. 

Created by: Otu Julia ujunwa 
Reg. No: 2024714031